- Print `errno` on socket errors
- Add error messages for all cases where error can happen
- Error handling in daemon
- Error handling in dexpo
- Handle xcb errors with `xcb_generic_error_t`

+++ src/socket.cpp
@@ -10,10 +10,12 @@
 /**
  * Thrower for custom errors.
  * Exists to make error throwing fit into one line.
+ *
+ * Its name looks readable: is<socket_error> (this->fd)
  */
 template <typename E, typename S>
 void
-is (S status) // This name looks readable: is<socket_error> (this->fd)
+is (S status, const std::string &error_msg)
 {
   // Check if this is one of the custom errors, because this
   // won't work for functions that return something other than -1 on error
@@ -21,7 +23,8 @@ is (S status) // This name looks readable: is<socket_error> (this->fd)
 
   if (status == -1) // Most of the C libs return -1 on error
     {
-      throw E ();
+      perror (E ().what ());
+      throw E (error_msg);
     }
 }
 
@@ -33,7 +36,7 @@ is (S status) // This name looks readable: is<socket_error> (this->fd)
  */
 template <typename D>
 void
-read_unix (int fd, D *dest, size_t length)
+read_unix (int fd, D *dest, size_t length, const std::string &error_msg)
 {
   size_t received = 0; // Number of bytes that we have already read
   ssize_t rcv = 0;     // Response of the read(2)
@@ -42,7 +45,7 @@ read_unix (int fd, D *dest, size_t length)
     {
       // Read what is left to read with an offset
       rcv = read (fd, dest + received, length - received);
-      is<read_error> (rcv);
+      is<read_error> (rcv, error_msg);
 
       // Increment the received amount with the number of bytes just received
       received += rcv;
@@ -57,7 +60,7 @@ read_unix (int fd, D *dest, size_t length)
  */
 template <typename D>
 void
-write_unix (int fd, D *src, size_t length)
+write_unix (int fd, D *src, size_t length, const std::string &error_msg)
 {
   size_t sent = 0; // Number of bytes that we have already sent
   ssize_t wr = 0;  // Response of the write(2). Number of bytes written
@@ -66,7 +69,7 @@ write_unix (int fd, D *src, size_t length)
     {
       // Send what is left to send with an offset
       wr = write (fd, src + sent, length - sent);
-      is<read_error> (wr);
+      is<write_error> (wr, error_msg);
 
       // Increment the sent amount with the number of bytes just transferred
       sent += wr;
@@ -101,37 +104,24 @@ dxp_socket::dxp_socket ()
   // XXX Make socket abstract
   // sock_name.sun_path[0] = '\0';
 
-  int s = 0; // Return status of functions to check for errors
-  try
-    {
-      // Create a socket file descriptor
-      this->fd = socket (AF_UNIX, SOCK_STREAM, 0);
-      is<socket_error> (this->fd);
+  int s = 0; ///< Return status of functions to check for errors
 
-      // Connect to named socket
-      s = connect (this->fd, sock_addr, sizeof (sock_name));
-      is<connect_error> (s);
-    }
-  catch (const connect_error &e)
+  // Create a socket file descriptor
+  this->fd = socket (AF_UNIX, SOCK_STREAM, 0);
+  is<socket_error> (this->fd, "Failed to create a socket file descriptor");
+
+  // Connect to named socket
+  s = connect (this->fd, sock_addr, sizeof (sock_name));
+  if (s == -1)
     {
-      // This should be run only for the first connection, as socket may not
-      // exist then.
-      // And should not be run when there are active connections to the socket.
+      // This should be run only for the initial connection,
+      // (daemon startup) as socket may not exist then.
 
       unlink (k_socket_path); // Remove existing socket
 
       s = bind (this->fd, sock_addr, sizeof (sock_name)); // Bind name to fd
-      s = listen (this->fd, 2);                           // 2 is arbitrary
-      is<bind_error> (s);
-    }
-  catch (const socket_error &e)
-    {
-      std::cerr << e.what () << std::endl;
-      throw;
-    }
-  catch (...)
-    {
-      throw;
+      s = listen (this->fd, 2);                           /* 2 is arbitrary */
+      is<bind_error> (s, "Failed to bind a name to the socket");
     }
 };
 
@@ -141,31 +131,35 @@ dxp_socket::~dxp_socket () { close (this->fd); };
  * Request and receive socket_pixmaps from daemon
  */
 std::vector<dxp_socket_desktop>
-dxp_socket::get_pixmaps () const
+dxp_socket::get_desktops () const
 {
   // Send pixmap request to daemon
-  char cmd = kRequestPixmaps;
-  write_unix (this->fd, &cmd, 1);
+  char cmd = kRequestDesktops;
+  write_unix (this->fd, &cmd, 1,
+              "Failed to send a desktops request to the daemon. "
+              "Please check if the daemon is running");
 
-  std::vector<dxp_socket_desktop>
-      pixmap_array; // Pixmap array that will be returned
+  /// Pixmap array that will be returned
+  std::vector<dxp_socket_desktop> pixmap_array;
 
-  uint num = 0; // Amount of pixmaps that will be received
-  read_unix (this->fd, &num, sizeof (num));
+  size_t num = 0; ///< Amount of desktops that will be received
+  read_unix (this->fd, &num, sizeof (num),
+             "Failed to get number of desktops from the daemon");
 
-  for (; num > 0; num--) // Read `num` pixmaps from socket
+  for (; num > 0; num--) // Read `num` desktops from socket
     {
-      // TODO(mmskv): Check for errors here
-      dxp_socket_desktop p;
+      dxp_socket_desktop p; ///< Desktop struct for read data
 
-      // Reading everything except raw pixmap
-      read_unix (this->fd, &p, offsetof (dxp_socket_desktop, pixmap));
+      // Reading all desktop data except raw pixmap
+      read_unix (this->fd, &p, offsetof (dxp_socket_desktop, pixmap),
+                 "Failed to get desktop data from the daemon");
 
-      // Allocating space for an incoming pixmap
+      // Allocating space for an incoming raw pixmap
       p.pixmap.resize (p.pixmap_len);
 
       // Reading raw pixmap
-      read_unix (this->fd, p.pixmap.data (), p.pixmap_len);
+      read_unix (this->fd, p.pixmap.data (), size_t (p.pixmap_len),
+                 "Failed to get raw desktop pixmap from the daemon");
 
       pixmap_array.push_back (p);
     }
@@ -173,13 +167,13 @@ dxp_socket::get_pixmaps () const
 };
 
 /**
- * Starts an infinite loop that listens for the kRequestPixmaps write
- * and sends pixmaps one by one in return
+ * Starts an infinite loop that listens for the kRequestDesktops write
+ * and sends desktops one by one in return
  */
 void
-dxp_socket::send_pixmaps_on_event (
-    const std::vector<dxp_socket_desktop> &pixmaps,
-    std::mutex &pixmaps_lock) const
+dxp_socket::send_desktops_on_event (
+    const std::vector<dxp_socket_desktop> &desktops,
+    std::mutex &desktops_lock) const
 {
   int data_fd = 0; // Socket file descriptor
 
@@ -189,26 +183,29 @@ dxp_socket::send_pixmaps_on_event (
     {
       // Check the incoming command
       char cmd = -1;
-      read_unix (data_fd, &cmd, 1);
+      read_unix (data_fd, &cmd, 1, "Failed to get a command from dexpo");
 
-      if (cmd == kRequestPixmaps)
+      if (cmd == kRequestDesktops)
         {
-          // Lock pixmaps to prevent race condition when reading and writing
-          // pixmaps to socket
-          std::scoped_lock<std::mutex> guard (pixmaps_lock);
+          // Lock desktops to prevent race condition when reading and writing
+          // desktops to socket
+          std::scoped_lock<std::mutex> guard (desktops_lock);
 
-          // First write -- number of subsequent packets
-          uint num = pixmaps.size ();
-          write_unix (data_fd, &num, sizeof (num));
+          // First write -- number of desktops to be sent
+          size_t num = desktops.size ();
+          write_unix (data_fd, &num, sizeof (num),
+                      "Failed to send number of desktops to dexpo");
 
-          // Writes from second to `num+1` -- subsequent packets
-          for (const auto &p : pixmaps)
+          // Writes from second to `num+1` * 2 -- sending desktops
+          for (const auto &p : desktops)
             {
               // Sending everything except raw pixmap
-              write_unix (data_fd, &p, offsetof (dxp_socket_desktop, pixmap));
+              write_unix (data_fd, &p, offsetof (dxp_socket_desktop, pixmap),
+                          "Failed to send desktop data to dexpo");
 
               // Sending pixmap
-              write_unix (data_fd, p.pixmap.data (), p.pixmap_len);
+              write_unix (data_fd, p.pixmap.data (), p.pixmap_len,
+                          "Failed to send raw desktop pixmap to dexpo");
             }
         }
     }
